+++ libs/artifact/CHANGELOG.rst
@@ -1,3 +1,8 @@
+**1.0.3 - 06/17/26**
+
+- Switch Jenkins ``vivarium_build_utils`` shared library loading from the ``epic/monorepo``
+  branch to the version returned by ``get_vbu_version()``
+
 **1.0.2 - 06/08/26**
 
 - Extract ``EntityKey`` into its own module and export ``is_entity_key`` helper